pdf_writer_service/requirements.txt
aiokafka==0.8.1
python-dotenv==1.0.0
reportlab==4.0.7

